Better FullName in examples
