+++ samples/ConfigSample/appsettings.json
@@ -27,20 +27,20 @@
       }
     },
     "Events": {
-      "ImageUploaded": { // should be the FullName of the type
+      "ConfigSample.ImageUploaded": { // FullName of the type
         "TransportName": "in-memory-images"
       },
-      "VideoUploaded": { // should be the FullName of the type
+      "ConfigSample.VideoUploaded": { // FullName of the type
         "TransportName": "in-memory-videos",
         "Consumers": {
-          "VisualsUploadedConsumer": { // should be the FullName of the type
+          "ConfigSample.VisualsUploadedConsumer": { // FullName of the type
             "UnhandledErrorBehaviour": "Discard",
             "Metadata": {
               "generation": "2022"
             }
           }
         }
-      },
+      }
     }
   }
 }
